+++ Assets/Scripts/Objects/Interfaces/IBreakable.cs
@@ -4,11 +4,11 @@ namespace Objects.Interfaces
 {
 	public interface IBreakable
 	{
-		public int Health { get; }
+		public float Health { get; }
 		
 		public bool IsBroken { get; }
 
-		public void Damage(int damage, object source);
+		public void Damage(float damage, object source);
 		public void Break(object source);
 
 		public GameObject GetGameObject();
